requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
openai>=1.59.0
chromadb>=1.5.0
PyMuPDF>=1.25.0
python-docx>=1.1.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
aiosqlite>=0.20.0
aiofiles>=24.1.0
python-dotenv>=1.0.1
tiktoken>=0.8.0
pyjwt>=2.8.0
bcrypt>=4.0.0
stripe>=11.0.0
